Chore(.jwt): move package
